Show settings menu in MainActivity
