+++ app/src/main/java/com/goobar/io/ad340/DailyForecastListAdapter.kt
@@ -9,11 +9,13 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 
-class DailyForecastViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+class DailyForecastViewHolder(
+    view: View,
+    private val tempDisplaySettingManager: TempDisplaySettingManager
+) : RecyclerView.ViewHolder(view) {
 
     private val tempText = view.findViewById<TextView>(R.id.tempText)
     private val descriptionText = view.findViewById<TextView>(R.id.descriptionText)
-    private val tempDisplaySettingManager = TempDisplaySettingManager(view.context)
 
     fun bind(dailyForecast: DailyForecast) {
         tempText.text = formatTempForDisplay(dailyForecast.temp, tempDisplaySettingManager.getTempDisplaySetting())
@@ -22,12 +24,13 @@ class DailyForecastViewHolder(view: View) : RecyclerView.ViewHolder(view) {
 }
 
 class DailyForecastListAdapter(
+    private val tempDisplaySettingManager: TempDisplaySettingManager,
     private val clickHandler: (DailyForecast) -> Unit
 ) : ListAdapter<DailyForecast, DailyForecastViewHolder>(DIFF_CONFIG) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): DailyForecastViewHolder {
         val itemView = LayoutInflater.from(parent.context).inflate(R.layout.item_daily_forecast, parent, false)
-        return DailyForecastViewHolder(itemView)
+        return DailyForecastViewHolder(itemView, tempDisplaySettingManager)
     }
 
     override fun onBindViewHolder(holder: DailyForecastViewHolder, position: Int) {
