+++ app/src/main/java/com/goobar/io/ad340/ForecastUtils.kt
@@ -1,5 +1,10 @@
 package com.goobar.io.ad340
 
+import android.content.Context
+import android.content.DialogInterface
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
+
 fun formatTempForDisplay(temp: Float, tempDisplaySetting: TempDisplaySetting) : String {
     return when (tempDisplaySetting) {
         TempDisplaySetting.Fahrenheit -> String.format("%.2f F°", temp)
@@ -8,4 +13,23 @@ fun formatTempForDisplay(temp: Float, tempDisplaySetting: TempDisplaySetting) : 
             String.format("%.2f C°", temp)
         }
     }
+}
+
+fun showTempDisplaySettingDialog(context: Context, tempDisplaySettingManager: TempDisplaySettingManager) {
+    val dialogBuilder = AlertDialog.Builder(context)
+        .setTitle("Choose Display Units")
+        .setMessage("Choose which temperature unit to use for temperature display")
+        .setPositiveButton("F°", object: DialogInterface.OnClickListener {
+            override fun onClick(dialog: DialogInterface?, which: Int) {
+                tempDisplaySettingManager.updateSetting(TempDisplaySetting.Fahrenheit)
+            }
+        })
+        .setNeutralButton("C°") { _,_ ->
+            tempDisplaySettingManager.updateSetting(TempDisplaySetting.Celsius)
+        }
+        .setOnDismissListener {
+            Toast.makeText(context, "Setting will take affect on app restart", Toast.LENGTH_SHORT).show()
+        }
+
+    dialogBuilder.show()
 }
