+++ app/src/main/java/com/goobar/io/ad340/MainActivity.kt
@@ -1,11 +1,16 @@
 package com.goobar.io.ad340
 
+import android.content.DialogInterface
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
 import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -14,12 +19,15 @@ import com.goobar.io.ad340.details.ForecastDetailsActivity
 class MainActivity : AppCompatActivity() {
 
   private val forecastRepository = ForecastRepository()
+  private lateinit var tempDisplaySettingManager: TempDisplaySettingManager
 
   // region Setup Methods
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     setContentView(R.layout.activity_main)
+
+    tempDisplaySettingManager = TempDisplaySettingManager(this)
 
     val zipcodeEditText: EditText = findViewById(R.id.zipcodeEditText)
     val enterButton: Button = findViewById(R.id.enterButton)
@@ -36,7 +44,7 @@ class MainActivity : AppCompatActivity() {
 
     val dailyForecastList: RecyclerView = findViewById(R.id.dailyForecastList)
     dailyForecastList.layoutManager = LinearLayoutManager(this)
-    val dailyForecastAdapter = DailyForecastListAdapter() {
+    val dailyForecastAdapter = DailyForecastListAdapter(tempDisplaySettingManager) {
       showForecastDetails(it)
     }
     dailyForecastList.adapter = dailyForecastAdapter
@@ -47,6 +55,23 @@ class MainActivity : AppCompatActivity() {
       dailyForecastAdapter.submitList(forecastItems)
     }
     forecastRepository.weeklyForecast.observe(this, weeklyForecastObserver)
+  }
+
+  override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+    val inflater: MenuInflater = menuInflater
+    inflater.inflate(R.menu.settings_menu, menu)
+    return true
+  }
+
+  override fun onOptionsItemSelected(item: MenuItem): Boolean {
+    // Handle item selection
+    return when (item.itemId) {
+      R.id.tempDisplaySetting -> {
+        showTempDisplaySettingDialog(this, tempDisplaySettingManager)
+        true
+      }
+      else -> super.onOptionsItemSelected(item)
+    }
   }
 
   private fun showForecastDetails(forecast: DailyForecast) {
